LPS-200994 SF, move package
